backend/requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# AI and ML
google-generativeai==0.8.3
python-dotenv==1.0.0
python-dateutil==2.9.0.post0

# Data validation and cleaning
pydantic==2.8.2
numpy==1.26.4
email-validator==2.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# File handling
aiofiles==23.2.1

# Data visualization (for server-side chart generation)
matplotlib==3.8.1
seaborn==0.13.0
plotly==5.17.0

# Text processing and NLP
nltk==3.8.1